Split software and hardware modules
Both have to be imported explicitly now.
