+++ raw-feldspar.cabal
@@ -19,15 +19,21 @@ source-repository head
 
 library
   exposed-modules:
-    Feldspar
     Data.VirtualContainer
     Feldspar.Representation
     Feldspar.Optimize
     Feldspar.Frontend
     Feldspar.Storable
+    Feldspar
+    Feldspar.Software.Representation
+    Feldspar.Software.Frontend
+    Feldspar.Software.Compile
+    Feldspar.Software
+    Feldspar.Hardware.Representation
+    Feldspar.Software.Frontend
+    Feldspar.Hardware.Compile
+    Feldspar.Hardware
     Feldspar.Vector
-    Feldspar.Compile.Software
-    Feldspar.Compile.Hardware
 
   other-modules:
     Data.VirtualContainer.TH
