+++ src/Feldspar/Frontend.hs
@@ -3,13 +3,8 @@
 module Feldspar.Frontend where
 
 
-import Prelude (Integral, error, reverse)
+import Prelude (Integral, error, (=<<), sequence_)
 import Prelude.EDSL
-
-import Control.Monad
-import Control.Monad.Trans
-
-import Data.Proxy
 
 import Language.Syntactic (Internal)
 import Language.Syntactic.Functional
@@ -17,17 +12,13 @@ import qualified Language.Syntactic as Syntactic
 
 import Language.Syntactic.TypeRep
 
-import Language.Embedded.Imperative.Frontend.General hiding (Ref, Arr)
-
-import qualified Control.Monad.Operational.Higher   as H
-
-import qualified Language.Embedded.Imperative       as Soft
-import qualified Language.Embedded.Imperative.CMD   as Soft
-
-import qualified Language.Embedded.Hardware as Hard
+import Language.Embedded.Imperative (IxRange)
+import qualified Language.Embedded.Imperative as Imp
 
 import Data.VirtualContainer
 import Feldspar.Representation
+
+
 
 --------------------------------------------------------------------------------
 -- * Pure expressions
@@ -138,7 +129,7 @@ max a b = a>=b ? a $ b
 unsafeArrIx :: Type a => Arr a -> Data Index -> Data a
 unsafeArrIx arr i = desugar $ mapVirtual arrIx $ unArr arr
   where
-    arrIx :: SmallType b => Soft.Arr Index b -> Data b
+    arrIx :: SmallType b => Imp.Arr Index b -> Data b
     arrIx arr = sugarSymTR (UnsafeArrIx arr) i
 
 
@@ -220,37 +211,37 @@ class Monad m => Controls m
 
 instance References Comp
   where
-    newRef   = fmap Ref $ mapVirtualA (const (Comp Soft.newRef)) virtRep
-    initRef  = fmap Ref . mapVirtualA (Comp . Soft.initRef) . sugar
-    getRef   = fmap desugar . mapVirtualA (Comp . Soft.getRef) . unRef
-    setRef r = sequence_ . zipListVirtual (\r' a' -> Comp $ Soft.setRef r' a') (unRef r) . sugar
+    newRef   = fmap Ref $ mapVirtualA (const (Comp Imp.newRef)) virtRep
+    initRef  = fmap Ref . mapVirtualA (Comp . Imp.initRef) . sugar
+    getRef   = fmap desugar . mapVirtualA (Comp . Imp.getRef) . unRef
+    setRef r = sequence_ . zipListVirtual (\r' a' -> Comp $ Imp.setRef r' a') (unRef r) . sugar
     modifyRef r f   = setRef r . f =<< unsafeFreezeRef r
-    unsafeFreezeRef = fmap desugar . mapVirtualA (Comp . Soft.unsafeFreezeRef) . unRef
+    unsafeFreezeRef = fmap desugar . mapVirtualA (Comp . Imp.unsafeFreezeRef) . unRef
 
 instance Arrays Comp
   where
     newArr :: forall a. Type a => Data Length -> Comp (Arr a)
-    newArr l = fmap Arr $ mapVirtualA (const (Comp $ Soft.newArr l)) rep
+    newArr l = fmap Arr $ mapVirtualA (const (Comp $ Imp.newArr l)) rep
       where rep = virtRep :: VirtualRep SmallType a
 
-    getArr i = fmap desugar . mapVirtualA (Comp . Soft.getArr i) . unArr
+    getArr i = fmap desugar . mapVirtualA (Comp . Imp.getArr i) . unArr
 
     setArr :: forall a. Type a => Data Index -> Data a -> Arr a -> Comp ()
-    setArr i a arr = sequence_ $ zipListVirtual (\a' arr' -> Comp $ Soft.setArr i a' arr') (rep) (unArr arr)
+    setArr i a arr = sequence_ $ zipListVirtual (\a' arr' -> Comp $ Imp.setArr i a' arr') (rep) (unArr arr)
       where rep = sugar a :: Virtual SmallType Data a
 
     copyArr arr1 arr2 len = sequence_ $
-        zipListVirtual (\a1 a2 -> Comp $ Soft.copyArr a1 a2 len)
+        zipListVirtual (\a1 a2 -> Comp $ Imp.copyArr a1 a2 len)
           (unArr arr1)
           (unArr arr2)
 
 instance Controls Comp
   where
-    iff c t f       = Comp $ Soft.iff c (unComp t) (unComp f)
-    for  range body = Comp $ Soft.for range (unComp . body)
-    while cont body = Comp $ Soft.while (unComp cont) (unComp body)
-    break           = Comp Soft.break
-    assert cond msg = Comp $ Soft.assert cond msg
+    iff c t f       = Comp $ Imp.iff c (unComp t) (unComp f)
+    for  range body = Comp $ Imp.for range (unComp . body)
+    while cont body = Comp $ Imp.while (unComp cont) (unComp body)
+    break           = Comp Imp.break
+    assert cond msg = Comp $ Imp.assert cond msg
 
 -- | Conditional statement that returns an expression
 ifE :: (References m, Controls m, Type a)
@@ -263,300 +254,3 @@ ifE c t f = do
     iff c (t >>= setRef res) (f >>= setRef res)
     unsafeFreezeRef res
 
-
-
---------------------------------------------------------------------------------
--- * Software specific operations.
---------------------------------------------------------------------------------
-
-liftS :: Comp a -> Software a
-liftS = Software . lift . unComp
-
-instance References (Software)
-  where
-    newRef          = liftS newRef
-    initRef         = liftS . initRef
-    getRef          = liftS . getRef
-    setRef r        = liftS . setRef r
-    modifyRef r     = liftS . modifyRef r
-    unsafeFreezeRef = liftS . unsafeFreezeRef
-
-instance Arrays (Software)
-  where
-    newArr        = liftS . newArr
-    getArr i      = liftS . getArr i
-    setArr i v    = liftS . setArr i v
-    copyArr a1 a2 = liftS . copyArr a1 a2
-
-instance Controls (Software)
-  where
-    iff c t f       = Software $ Soft.iff c (unSoftware t) (unSoftware f)
-    for  range body = Software $ Soft.for range (unSoftware . body)
-    while cont body = Software $ Soft.while (unSoftware cont) (unSoftware body)
-    break           = Software Soft.break
-    assert cond msg = Software $ Soft.assert cond msg
-
-
-
-----------------------------------------
--- ** Pointer operations
-----------------------------------------
-
--- | Swap two pointers
---
--- This is generally an unsafe operation. E.g. it can be used to make a
--- reference to a data structure escape the scope of the data.
---
--- The 'IsPointer' class ensures that the operation is only possible for types
--- that are represented as pointers in C.
-unsafeSwap :: IsPointer a => a -> a -> Software ()
-unsafeSwap a b = Software $ Soft.unsafeSwap a b
-
-
-
-----------------------------------------
--- ** File handling
-----------------------------------------
-
--- | Open a file
-fopen :: FilePath -> IOMode -> Software Handle
-fopen file = Software . Soft.fopen file
-
--- | Close a file
-fclose :: Handle -> Software ()
-fclose = Software . Soft.fclose
-
--- | Check for end of file
-feof :: Handle -> Software (Data Bool)
-feof = Software . Soft.feof
-
-class PrintfType r
-  where
-    fprf :: Handle -> String -> [Soft.PrintfArg Data] -> r
-
-instance (a ~ ()) => PrintfType (Software a)
-  where
-    fprf h form = Software . H.singleE . Soft.FPrintf h form . reverse
-
-instance (Formattable a, SmallType a, PrintfType r) => PrintfType (Data a -> r)
-  where
-    fprf h form as = \a -> fprf h form (Soft.PrintfArg a : as)
-
--- | Print to a handle. Accepts a variable number of arguments.
-fprintf :: PrintfType r => Handle -> String -> r
-fprintf h format = fprf h format []
-
--- | Put a single value to a handle
-fput :: (Formattable a, SmallType a)
-    => Handle
-    -> String  -- Prefix
-    -> Data a  -- Expression to print
-    -> String  -- Suffix
-    -> Software ()
-fput h pre e post = Software $ Soft.fput h pre e post
-
--- | Get a single value from a handle
-fget :: (Formattable a, SmallType a) => Handle -> Software (Data a)
-fget = Software . Soft.fget
-
--- | Print to @stdout@. Accepts a variable number of arguments.
-printf :: PrintfType r => String -> r
-printf = fprintf Soft.stdout
-
-
-
-----------------------------------------
--- ** Abstract objects
-----------------------------------------
-
-newObject
-    :: String  -- ^ Object type
-    -> Software Object
-newObject = Software . Soft.newObject
-
-initObject
-    :: String        -- ^ Function name
-    -> String        -- ^ Object type
-    -> [FunArg Data] -- ^ Arguments
-    -> Software Object
-initObject fun ty args = Software $ Soft.initObject fun ty args
-
-initUObject
-    :: String        -- ^ Function name
-    -> String        -- ^ Object type
-    -> [FunArg Data] -- ^ Arguments
-    -> Software Object
-initUObject fun ty args = Software $ Soft.initUObject fun ty args
-
-
-
-----------------------------------------
--- ** External function calls (C-specific)
-----------------------------------------
-
--- | Add an @#include@ statement to the generated code
-addInclude :: String -> Software ()
-addInclude = Software . Soft.addInclude
-
--- | Add a global definition to the generated code
---
--- Can be used conveniently as follows:
---
--- > {-# LANGUAGE QuasiQuotes #-}
--- >
--- > import Feldspar.IO
--- >
--- > prog = do
--- >     ...
--- >     addDefinition myCFunction
--- >     ...
--- >   where
--- >     myCFunction = [cedecl|
--- >       void my_C_function( ... )
--- >       {
--- >           // C code
--- >           // goes here
--- >       }
--- >       |]
-addDefinition :: Soft.Definition -> Software ()
-addDefinition = Software . Soft.addDefinition
-
--- | Declare an external function
-addExternFun :: forall proxy res . SmallType res
-    => String         -- ^ Function name
-    -> proxy res      -- ^ Proxy for expression and result type
-    -> [FunArg Data]  -- ^ Arguments (only used to determine types)
-    -> Software ()
-addExternFun fun res args = Software $ Soft.addExternFun fun res' args
-  where
-    res' = Proxy :: Proxy (Data res)
-
--- | Declare an external procedure
-addExternProc
-    :: String         -- ^ Procedure name
-    -> [FunArg Data]  -- ^ Arguments (only used to determine types)
-    -> Software ()
-addExternProc proc args = Software $ Soft.addExternProc proc args
-
--- | Call a function
-callFun :: SmallType a
-    => String         -- ^ Function name
-    -> [FunArg Data]  -- ^ Arguments
-    -> Software (Data a)
-callFun fun as = Software $ Soft.callFun fun as
-
--- | Call a procedure
-callProc
-    :: String         -- ^ Function name
-    -> [FunArg Data]  -- ^ Arguments
-    -> Software ()
-callProc fun as = Software $ Soft.callProc fun as
-
--- | Declare and call an external function
-externFun :: SmallType res
-    => String         -- ^ Procedure name
-    -> [FunArg Data]  -- ^ Arguments
-    -> Software (Data res)
-externFun fun args = Software $ Soft.externFun fun args
-
--- | Declare and call an external procedure
-externProc
-    :: String         -- ^ Procedure name
-    -> [FunArg Data]  -- ^ Arguments
-    -> Software ()
-externProc proc args = Software $ Soft.externProc proc args
-
--- | Get current time as number of seconds passed today
-getTime :: Software (Data Double)
-getTime = Software Soft.getTime
-
--- | Constant string argument
-strArg :: String -> FunArg Data
-strArg = Soft.strArg
-
--- | Value argument
-valArg :: SmallType a => Data a -> FunArg Data
-valArg = Soft.valArg
-
--- | Reference argument
-refArg :: SmallType a => Ref a -> FunArg Data
-refArg (Ref r) = Soft.refArg (viewActual r)
-
--- | Array argument
-arrArg :: SmallType a => Arr a -> FunArg Data
-arrArg (Arr a) = Soft.arrArg (viewActual a)
-
--- | Abstract object argument
-objArg :: Object -> FunArg Data
-objArg = Soft.objArg
-
--- | Modifier that takes the address of another argument
-addr :: FunArg Data -> FunArg Data
-addr = Soft.addr
-
-
-
---------------------------------------------------------------------------------
--- * Hardware specific operations.
---------------------------------------------------------------------------------
-
---------------------------------------------------------------------------------
--- ** Signals.
-
-newtype Sig a = Sig { unSig :: Virtual SmallType Hard.Signal a }
-
--- | Create an uninitialized signal.
-newSig :: Type a => Hardware (Sig a)
-newSig = fmap Sig $ mapVirtualA (const (Hardware Hard.newSignal_)) virtRep
-
--- | Create an initialized signal.
-initSig :: Type a => Data a -> Hardware (Sig a)
-initSig = fmap Sig . mapVirtualA (Hardware . Hard.newSignal) . sugar
-
--- | Get the contents of a signal.
-getSig :: Type a => Sig a -> Hardware (Data a)
-getSig = fmap desugar . mapVirtualA (Hardware . Hard.getSignal) . unSig
-
--- | Set the contents of a signal.
-setSig :: Type a => Sig a -> Data a -> Hardware ()
-setSig s = sequence_ . zipListVirtual (\s' a' -> Hardware $ Hard.setSignal s' a') (unSig s) . sugar
-
--- | Modify the contents of a signal.
-modifySig :: Type a => Sig a -> (Data a -> Data a) -> Hardware ()
-modifySig s f = setSig s . f =<< unsafeFreezeSig s
-
--- | Freeze the contents of a signal.
-unsafeFreezeSig :: Type a => Sig a -> Hardware (Data a)
-unsafeFreezeSig = fmap desugar . mapVirtualA (Hardware . Hard.unsafeFreezeSignal) . unSig
-
---------------------------------------------------------------------------------
--- ** ...
-
-liftH :: Comp a -> Hardware a
-liftH = Hardware . lift . unComp
-
-instance References (Hardware)
-  where
-    newRef          = liftH newRef
-    initRef         = liftH . initRef
-    getRef          = liftH . getRef
-    setRef r        = liftH . setRef r
-    modifyRef r     = liftH . modifyRef r
-    unsafeFreezeRef = liftH . unsafeFreezeRef
-
-instance Arrays (Hardware)
-  where
-    newArr        = liftH . newArr
-    getArr i      = liftH . getArr i
-    setArr i v    = liftH . setArr i v
-    copyArr a1 a2 = liftH . copyArr a1 a2
-
-instance Controls (Hardware)
-  where
-    iff c t f          = Hardware $ Hard.iff c (unHardware t) (unHardware f)
-    for (i, _, _) body = Hardware $ Hard.for i (unHardware . body)
-    while cont body    = Hardware $ Hard.while (unHardware cont) (unHardware body)
-    break              = liftH break
-    assert cond msg    = liftH $ assert cond msg
-
---------------------------------------------------------------------------------
