+++ src/Feldspar/Hardware.hs
@@ -1,0 +1,17 @@
+module Feldspar.Hardware
+  ( -- * Front end
+    module Feldspar.Hardware.Frontend
+    -- * Back ends
+  , runIO
+  , compile
+  , icompile
+  , compile2
+  , icompile2
+  ) where
+
+
+
+import Feldspar.Hardware.Representation
+import Feldspar.Hardware.Compile
+import Feldspar.Hardware.Frontend
+
